Fix: Improved logic, names format and initial accumulator value in reduce function
