+++ 1-js-and-async-programming/2-asynchronous-js/3-async-await.js
@@ -10,9 +10,26 @@
     - Use await with the methods from utils/mocked-api to get the data
     - Make sure to return a string containing the name of the most common subscription
 */
-
-const { getUsersWithMoreDislikedMoviesThanLikedMoviesAsync } = require("./2-promise");
+const { getUsers, getLikedMovies, getDislikedMovies } = require("./utils/mocked-api");
 const { getUserSubscriptionByUserId } = require("./utils/mocked-api");
+
+const getUsersWithMoreDislikedMoviesThanLikedMoviesAsync = async () => {
+  try {
+    const users = await getUsers();
+    const likedMovies = await getLikedMovies();
+    const dislikedMovies = await getDislikedMovies();
+    
+    const filteredUsers = await Promise.all(
+      users.filter(async user => {
+        let likedMoviesCount = likedMovies.find(m => m.userId == user.id)?.movies.length || 0;
+        let dislikedMoviesCount = dislikedMovies.find(m => m.userId == user.id)?.movies.length || 0;
+        return dislikedMoviesCount > likedMoviesCount;
+      }));
+    return filteredUsers;
+  } catch (error) {
+    throw error;
+  }
+};
 
 /**
  * Logs the most common subscription among users
@@ -27,12 +44,13 @@ const getCommonDislikedSubscription = async () => {
     const suscriptions = await Promise.all(
       harshReviewers.map(async reviewer => await getUserSubscriptionByUserId(reviewer.id)
       ));
-
-    const mostCommon = suscriptions.reduce((subs, sub) => {
-      subs[sub] = (subs[sub] || 0) + 1;
-      return subs;
-    });
-    return mostCommon.subscription;
+      
+      const subsCount = suscriptions.reduce((accumulator, item) => {
+        accumulator[item.subscription] = (accumulator[item.subscription] || 0) + 1;
+        return accumulator;
+      }, []);
+      const sortedSubsCountList = Object.entries(subsCount).sort((curr, next) => next[1] - curr[1]);
+    return sortedSubsCountList[0][0];
   } catch (error) {
     throw error;
   }
